JeffR: isolate generic lib, part II: split code w/ local proxies for start() vs lib.Start()
